Update and cleanup tags/about modal
