+++ client/src/queryEditor/toolbar/AboutContent.js
@@ -1,5 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
+import OpenInNewIcon from 'mdi-react/OpenInNewIcon';
+
+const LINK_STYLE = { display: 'inline-flex', alignItems: 'center' };
 
 function AboutContent({ version }) {
   return (
@@ -12,46 +15,68 @@ function AboutContent({ version }) {
           <strong>Update available</strong>: {version && version.latest}
         </p>
       )}
-      <p>
-        <strong>Project Page</strong>:{' '}
+
+      <div
+        style={{
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'space-around',
+          marginBottom: 16
+        }}
+      >
         <a
           href="http://rickbergfalk.github.io/sqlpad/"
           target="_blank"
           rel="noopener noreferrer"
+          style={LINK_STYLE}
         >
-          http://rickbergfalk.github.io/sqlpad
+          Project page <OpenInNewIcon size={18} />
         </a>
+        <a
+          href="https://github.com/rickbergfalk/sqlpad/issues"
+          target="_blank"
+          rel="noopener noreferrer"
+          style={LINK_STYLE}
+        >
+          Submit an Issue <OpenInNewIcon size={18} />
+        </a>
+        <a
+          href="https://github.com/rickbergfalk/sqlpad/blob/master/CHANGELOG.md"
+          target="_blank"
+          rel="noopener noreferrer"
+          style={LINK_STYLE}
+        >
+          Changelog <OpenInNewIcon size={18} />
+        </a>
+        <a
+          href="https://github.com/rickbergfalk/sqlpad"
+          target="_blank"
+          rel="noopener noreferrer"
+          style={LINK_STYLE}
+        >
+          GitHub <OpenInNewIcon size={18} />
+        </a>
+      </div>
+
+      <p>
+        <strong>Shortcuts</strong>
       </p>
-      <hr />
-      <ul>
-        <li role="presentation">
-          <a
-            href="https://github.com/rickbergfalk/sqlpad/issues"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Submit an Issue
-          </a>
+      <ul style={{ paddingLeft: 0 }}>
+        <li style={{ listStyleType: 'none', marginBottom: 8 }}>
+          <code>ctrl+s</code> / <code>command+s</code> : Save
         </li>
-        <li role="presentation">
-          <a
-            href="https://github.com/rickbergfalk/sqlpad/blob/master/CHANGELOG.md"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Changelog
-          </a>
+        <li style={{ listStyleType: 'none', marginBottom: 8 }}>
+          <code>ctrl+return</code> / <code>command+return</code> : Run
         </li>
-        <li role="presentation">
-          <a
-            href="https://github.com/rickbergfalk/sqlpad"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            GitHub Repository
-          </a>
+        <li style={{ listStyleType: 'none', marginBottom: 8 }}>
+          <code>shift+return</code> : Format
         </li>
       </ul>
+
+      <p>
+        <strong>Tip</strong>
+      </p>
+      <p>Run only a portion of a query by highlighting it first.</p>
     </div>
   );
 }
